requirements.txt
streamlit
google-cloud-storage
google-auth
google-generativeai
lxml